+++ lib/preconditions.dart
@@ -10,5 +10,4 @@ import 'package:flutter/widgets.dart';
 import 'package:logging/logging.dart';
 
 part 'src/preconditions_impl.dart';
-part 'src/preconditions_scope.dart';
 part 'src/preconditions_status.dart';